retire DATA_MODEL_ID setting
We cannot import from atlas.conf in a project-level settings file
due to circular import issues.

The intent of DATA_MODEL_ID was to provide a value that could be used
by site developers to customize when an ATLAS database would be re-used
_OR_ re-created.

Moving to a special data_model.VERSION constant allows the value to be
imported / inspected within settings.py.