refactor version importer
We had already bypassed most of django-treebeard's tree handlers
(add_child, add_sibling) in favor of the bulk_create manager.

This worked well if the majority of the nodes in a tree were
"text part level", but performance for "work level" nodes was subpar,
as those nodes were still resulting in multiple insert / select / update
SQL statements when created.

The importer now builds the entire tree in memory, and does a bulk insert
of the tree.  The tree has depth and path information.

The only thing missing at this time is the `numchild` value.
This could be calculated from the tree, but for now, I've chosen to
just override the method that we end up using (`get_descendants`).