Backport resolver fix from SV 1 fork
A few parts to this fix:

1) MyCapytain was never returning the text group metadata from the
inventory, so we were always working with the _last_ metadata object
found traversing the data directory.

2) After fixing MyCapytain, `process_text_group` was able to _update_
the text group metadata in the inventory. However, exceptions still occurred
because we were passing a different `XmlCtsTextgroupMetadata` object down to
`process_work`.

By updating the inventory and passing the _same_ `XmlCtsTextgroupMetadata`
object to `process_work`, the work URNs were available from both the
metadata object _and_ the inventory, causing ingestion to proceed as
expected.

3) We also need to ensure that `XmlCtsWorkMetadata` is updated if two different
corpora contribute texts to a particular work.