Initial pass at partial ingestion
We have to work around some of the "bulkification" present in CTSImporter.

Bulkification is really important for Brill (where there are textgroups with ~2.4k to 104k works.

I hope to circle back and test this branch out with Brill's content to see if there are more gains to be had, but this _should_ help within Beyond Translation, which was the original goal.